Add chamfer parameters for bin magnet cutout
To make friction fit magnet insertion more forgiving. This adds two chamfer parameters to the bin magnet cutout. In they UI they are:

Magnet cutout chamfer XY
Magnet cutout chamfer Z

If either is set to 0.0, which is the default, then the chamfer is turned off. I use 0.2 mm for XY and 1.0 for Z.